\name{CO2data}
\docType{data}
\title{GNP and CO2 Data Set}
\alias{CO2data}
\usage{
data(CO2data)
}
\description{
  This data set gives the gross national product (GNP) per capita in 1996 for various countries as
  well as their estimated carbon dioxide (CO2) emission per capita for the same year. 
}
\format{This data frame consists of 28 countries and the following columns:
\itemize{
  \item{\code{GNP}}{The gross national product per capita in 1996.}
  \item{\code{CO2}}{The estimated carbon dioxide emission per capita in 1996.}
  \item{\code{country}}{An abbreviation pertaining to the country measured 
  (e.g., "GRC" = Greece and "CH" = Switzerland).}
}
}
\references{
  Hurn, M., Justel, A. and Robert, C. P. (2003) Estimating Mixtures of Regressions, \emph{Journal 
  of Computational and Graphical Statistics} \bold{12(1)}, 55--79.
}
\keyword{datasets}
