\name{ddirichlet}
\alias{ddirichlet}

\title{Density Function for the Dirichlet Distribution}
\description{
  Density function for the Dirichlet distribution.
}
\usage{
ddirichlet(x, alpha)
}

\arguments{
  \item{x}{A k-dimensional vector of values that sum to 1 for which to calculate the density}
  \item{alpha}{A k-dimensional vector of the Dirichlet distribution parameters.}
}

\details{
  This is usually not to be called by the user.
}

\keyword{internal}
