\name{NOdata}
\docType{data}
\alias{NOdata}
\title{Ethanol Fuel Data Set}
\description{
  This data set gives the equivalence ratios and peak nitrogen oxide emissions in a
  study using pure ethanol as a spark-ignition engine fuel.  
}
\usage{
data(NOdata)
}
\format{This data frame consists of:
\itemize{
  \item{\code{NO}}{The peak nitrogen oxide emission levels.}
  \item{\code{Equivalence}}{The equivalence ratios for the engine at compression ratios from 7.5 to 18.}
}
}
\source{Brinkman, N. D. (1981) Ethanol Fuel -- A Single-Cylinder Engine Study of Efficiency and
Exhaust Emissions, \emph{S.A.E. Transactions}, 68.
}
\references{
  Hurn, M., Justel, A. and Robert, C. P. (2003) Estimating Mixtures of Regressions, \emph{Journal 
  of Computational and Graphical Statistics} \bold{12(1)}, 55--79.
}
\keyword{datasets}
