\name{parse.constraints}
\title{Constraint Function}
\alias{parse.constraints}
\usage{
parse.constraints(constr, k = 2, allsame = FALSE)
}
\description{
Constraint function for some mixture EM algorithms.
}
\arguments{
  \item{constr}{Vector indicating constrained/unconstrained means.}
  \item{k}{Number of components.}
  \item{allsame}{Logical indicating for processing the constraints.}
}
\details{
This function is not intended to be called by the user.
}


\keyword{internal}
