\name{RanEffdata}
\docType{data}
\title{Simulated Data from 2-Component Mixture of Regressions with Random Effects}
\alias{RanEffdata}
\usage{
data(RanEffdata)
}
\description{
  This data set was generated from a 2-component mixture of regressions with random effects. 
} 
\format{This data set consists of a list with 100 25x3 matrices. 
    The first column is the response variable, the second column is a column of 1's and the last column is the
    predictor variable.
}
\seealso{
\code{\link{regmixEM.mixed}}
}
\keyword{datasets}
