\name{RodFramedata}
\docType{data}
\title{Rod and Frame Task Data Set}
\alias{RodFramedata}
\usage{
data(RodFramedata)
}
\description{
  This data set involves assessing children longitudinally at 6 age points from ages 4 through 18 years for the rod and
  frame task.  This task sits the child in a darkened room in front of a luminous square frame tilted at 28 degrees on its
  axis to the left or right.  Centered inside the frame was a luminous rod also tilted 28 degrees to the left or right.  The
  child's task was to adjust the rod to the vertical position and the absolute deviation from the vertical (in degrees) was the
  measured response.
}
\format{This data frame consists of 140 children (the rows).  Column 1 is the subject number and column 
2 is the sex (0=MALE and 1=FEMALE).  Columns 3 through 26 give the 8 responses at each of the ages 4, 5, and 7.  
Columns 27 through 56 give the 10 responses at each of the ages 11, 14, and 18.  A value of 99 denotes missing data.
}
\source{
  Thomas, H. and Dahlin, M. P. (2005) Individual Development and Latent Groups: Analytical Tools for Interpreting Heterogeneity,
  \emph{Developmental Review} \bold{25(2)}, 133--154.
}
\keyword{datasets}
