\name{RTdata}
\docType{data}
\title{Reaction Time (RT) Data Set}
\alias{RTdata}
\usage{
data(RTdata)
}
\description{
  This data set involves normally developing children 9 years of age presented with two visual simuli on a computer monitor.
  The left image is the target stimuli and on the right is either an exact copy or a mirror image
  of the target stimuli.  The child must press one key if it is a copy or another key if it is a mirror image.  The
  data consists of the reaction times (RT) of the 197 children who provided correct responses for all 6 task trials. 
}
\format{This data frame consists of 197 children (the rows) and their 6 responses (the columns)
to the stimulus presented.  The response (RT) is recorded in milliseconds.
}
\references{
  Cruz-Medina, I. R., Hettmansperger, T. P. and Thomas, H. (2004) Semiparametric Mixture Models and Repeated
  Measures: The Multinomial Cut Point Model, \emph{Applied Statistics} \bold{53(3)}, 463--474.
  
  Miller, C. A., Kail, R., Leonard, L. B. and Tomblin, J. B. (2001) Speed of Processing in Children with
  Specific Language Impairment, \emph{Journal of Speech, Language, and Hearing Research} \bold{44(2)}, 416--433.
}
\seealso{
\code{\link{RTdata2}}
}
\keyword{datasets}
