\name{Waterdata}
\docType{data}
\title{Water-Level Task Data Set}
\alias{Waterdata}
\usage{Waterdata}
\description{
  This data set arises from the water-level task proposed by the Swiss psychologist Jean Piaget to
  assess children's understanding of the physical world.  This involves presenting a child with a 
  rectangular shaped, two-dimensional vessel on a sheet of paper tilted in specified clock-hour orientations.
  The child is then required to draw a line representing the still liquid in the tilted vessel.  Then,
  the degree of deviation about the horizontal is recorded. 
}
\format{This data frame consists of 405 children (the rows) and the degree of deviation about the horizontal
(the columns) for 8 specified clock-hour orientations (1, 2, 4, 5, 7, 8, 10, and 11 o'clock).
}
\source{
  Thomas, H. and Lohaus, A.  (1993)
  \emph{Modeling Growth and Individual Differences in Spatial Tasks},
  University of Chicago Press, Chicago.
} 
\keyword{datasets}
