\name{Waterdata}
\docType{data}
\title{Water-Level Task Data Set}
\alias{Waterdata}
\usage{Waterdata}
\description{
  This data set arises from the water-level task proposed by the Swiss psychologist Jean Piaget to
  assess children's understanding of the physical world.  This involves presenting a child with a 
  simple sketch of a rectangular vessel with a cap
  on a sheet of paper and tilted in specific orientations.
  The child is then required to draw a line representing the still liquid in the tilted vessel.  
  This line defines two points of intersection with the sides of the vessel.  A line
  through those points is drawn, then the acute angle between this line and the 
  horizontal (the correct orientation) is measured, with sign according to the slope
  of the line.  These signed angles are given, in degrees.
}
\format{This data frame consists of 405 children (the rows) and the degree of deviation about the horizontal
(the columns) for 8 specified clock-hour orientations 
(11, 4, 2, 7, 10, 5, 1, and 8 o'clock, in order).
}
\source{
  Thomas, H. and Lohaus, A.  (1993)
  \emph{Modeling Growth and Individual Differences in Spatial Tasks},
  University of Chicago Press, Chicago, available on JSTOR.
} 
\keyword{datasets}
